requirements.txt
asgiref==3.8.1
Django==5.1
django-environ==0.11.2
pillow==10.4.0
pip==24.0
psycopg==3.2.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
django-phonenumber-field==8.0.0
phonenumbers==8.13.22